requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
dnspython==2.4.2
idna==3.4
paho-mqtt==1.6.1
pymongo==4.5.0
python-decouple==3.8
requests==2.31.0
urllib3==2.0.5
